% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoggsaveR_functions.R
\name{auto_save_plot}
\alias{auto_save_plot}
\title{Automatically save a list of ggplot objects as a single image}
\usage{
auto_save_plot(
  plot_lst,
  relative_output_dir,
  file_name,
  ncol = 1,
  base_size = 20,
  verbose = TRUE
)
}
\arguments{
\item{plot_lst}{A list of ggplot objects}

\item{relative_output_dir}{relative output directory for the image to be saved}

\item{file_name}{The file name of the image to be saved}

\item{ncol}{The number of columns in the final plot (default is 1)}

\item{base_size}{The base size for the final plot (default is 20)}

\item{verbose}{Controls verbosity of output details}
}
\value{
This function does not return a value. It saves the final plot as an image.
}
\description{
This function takes a list of ggplot objects, combines them into a single
patchwork plot, calculates the optimal dimensions for the final plot based on
the number of plots, layers, facets, and the maximum number of items on the x
and y axes, and saves the final plot as an image.
}
\details{
Supports multiple plots in a single ggplot2 object as created by \code{library(patchwork)}
}
