% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoggsaveR_functions.R
\name{get_aspect_ratio}
\alias{get_aspect_ratio}
\title{This function calculates the aspect ratio by considering the number of facets and the range of the x and y axes.}
\usage{
get_aspect_ratio(plot_lst, plot_info, axes_info)
}
\arguments{
\item{plot_lst}{A list of ggplot objects}

\item{plot_info}{A list containing the plot information from ggplot_build}

\item{axes_info}{The maximum number of x and y elements in the list of plots}
}
\value{
The adjusted aspect ratio
}
\description{
This function first sets a base aspect ratio of 1.0.
It then adjusts this base ratio based on the number of facets
and the number of items on the x and y axes. If there are
multiple facets, the aspect ratio is increased, which makes
the plot wider. If there are more items on the x-axis than on
the y-axis, the aspect ratio is increased, which also makes the plot wider.
}
\details{
Conversely, if there are more items on the y-axis than on the x-axis,
the aspect ratio is decreased, which makes the plot taller.
}
