% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoggsaveR_functions.R
\name{get_axes_info}
\alias{get_axes_info}
\title{Get the number of items on the x and y axes of a ggplot object}
\usage{
get_axes_info(plot_lst)
}
\arguments{
\item{plot_lst}{A list of ggplot objects}
}
\value{
A list with the length or range of the x and y axes
}
\description{
This function extracts the data from a ggplot object and counts the unique number of
values of the x and y variables
}
