% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoggsaveR_functions.R
\name{get_num_plot_items}
\alias{get_num_plot_items}
\title{Get the number of items on the x and y axes of a ggplot object}
\usage{
get_num_plot_items(plot)
}
\arguments{
\item{plot}{A ggplot object}
}
\value{
A list with the length or range of the x and y axes, depending on whether they are continuous or discrete
}
\description{
This function extracts the data from a ggplot object and counts the unique
values of the x and y variables if it's discrete, or calculates the range if it's continuous.
It also calculates the range of the y variable.
}
