% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoggsaveR_functions.R
\name{get_plot_info}
\alias{get_plot_info}
\title{Get information about a list of plots}
\usage{
get_plot_info(plot_lst, verbose = FALSE)
}
\arguments{
\item{plot_lst}{A list of ggplot objects}
}
\value{
A list with the number of plots, layers, and facets
}
\description{
This function extracts the number of plots, the number of layers across all plots, and the number of facets
}
