% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoggsaveR_functions.R
\name{auto_save_plot}
\alias{auto_save_plot}
\title{Automatically save a list of ggplot objects as a single image}
\usage{
auto_save_plot(plot_lst, filename, ncol = 1, verbose = FALSE)
}
\arguments{
\item{plot_lst}{A list of ggplot objects}

\item{filename}{The file name of the image to be saved}

\item{ncol}{The number of columns for plots to be placed into in desired in the final image}

\item{verbose}{Controls verbosity of output details}
}
\value{
This function does not return a value. It saves the final plot as an image.
}
\description{
This function takes a list of ggplot objects, combines them into a single
patchwork plot, calculates the optimal dimensions for the final plot based on
the number of plots, layers, facets, and the maximum number of items on the x
and y axes, and saves the final plot as an image.
}
\details{
Supports multiple plots in a single ggplot2 object as created by \code{library(patchwork)}
}
