% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoggsaveR_functions.R
\name{get_plot_complexity}
\alias{get_plot_complexity}
\title{Calculate plot complexity}
\usage{
get_plot_complexity(base_size = 20, plot_info, ncol)
}
\arguments{
\item{base_size}{The base size for the final plot (default is 20)}

\item{plot_info}{The list of plot information as calculated by \code{get_plot_info}}

\item{ncol}{The number of columns in the output}
}
\value{
This function returns a number representing the complexity of the plot
}
\description{
This function calculates complexity of the final plot
based off of the number of plots, layers, and facets, and the
desired number of columns of plots, in the final plot
}
